+++ src/main/java/com/analytique/gestion_analytique/Repositories/PosteRepository.java
@@ -2,6 +2,6 @@ package com.analytique.gestion_analytique.Repositories;
 import com.analytique.gestion_analytique.Models.*;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PosteRepository extends JpaRepository<Poste, Long>{
+public interface PosteRepository extends JpaRepository<Poste, Integer>{
 
 }
