+++ src/main/java/com/analytique/gestion_analytique/Repositories/TypeContratRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.analytique.gestion_analytique.Models.TypeContrat;
 
-public interface TypeContratRepository extends JpaRepository<TypeContrat, Long>{
+public interface TypeContratRepository extends JpaRepository<TypeContrat, Integer>{
 
 }
