+++ src/main/java/com/analytique/gestion_analytique/controller/CompetenceContoller.java
@@ -28,7 +28,7 @@ public class CompetenceContoller {
 	}
 
 	@GetMapping("/poste/{id}")
-	public List<Competence> getCompetencesByPoste(@PathVariable Long id) {
+	public List<Competence> getCompetencesByPoste(@PathVariable Integer id) {
 		return cRepo.findByPoste(id);
 	}
